Move pending txs to a context to have identity related tx to be displayed (#652)
