+++ packages/ui/src/components/CallInfo.tsx
@@ -9,7 +9,7 @@ import { HiOutlineArrowTopRightOnSquare as LaunchIcon } from 'react-icons/hi2'
 import { Link } from './library'
 import { usePjsLinks } from '../hooks/usePjsLinks'
 import { Alert } from '@mui/material'
-import { CallDataInfoFromChain } from '../hooks/usePendingTx'
+import { CallDataInfoFromChain } from '../contexts/PendingTxContext'
 import { JSONprint } from '../utils/jsonPrint'
 import { Transaction } from 'polkadot-api'
 import MultisigCompactDisplay from './MultisigCompactDisplay'
