+++ packages/ui/src/components/modals/ChangeMultisig.tsx
@@ -28,7 +28,7 @@ import { MdErrorOutline as ErrorOutlineIcon } from 'react-icons/md'
 import { Button } from '../library'
 import { ModalCloseButton } from '../library/ModalCloseButton'
 import { useGetSortAddress } from '../../hooks/useGetSortAddress'
-import { useGetMultisigAddress } from '../../contexts/useGetMultisigAddress'
+import { useGetMultisigAddress } from '../../hooks/useGetMultisigAddress'
 import { getAsMultiTx } from '../../utils/getAsMultiTx'
 import { Enum } from 'polkadot-api'
 import { MultiAddress } from '@polkadot-api/descriptors'
