+++ packages/ui/src/components/modals/ProposalSigning.tsx
@@ -21,7 +21,7 @@ import { ModalCloseButton } from '../library/ModalCloseButton'
 import { useGetSortAddress } from '../../hooks/useGetSortAddress'
 import { useCheckTransferableBalance } from '../../hooks/useCheckTransferableBalance'
 import { getAsMultiTx } from '../../utils/getAsMultiTx'
-import { CallDataInfoFromChain } from '../../hooks/usePendingTx'
+import { CallDataInfoFromChain } from '../../contexts/PendingTxContext'
 import { debounce } from '../../utils/debounce'
 import { FixedSizeBinary, HexString, Transaction } from 'polkadot-api'
 import { useAnyApi } from '../../hooks/useAnyApi'
